split storage folder shrinking into its own file
